switching from boolean to thread-safe threading event, removing duplicate reporting, adding comments to explain service handling, improving safety and service handling, overriding pywin32util SvcRun so we no longer report as running before child processes are spawned

+++ agent/ncpa.py
@@ -86,6 +86,8 @@ if os.name == 'nt':
     import win32event
     import win32service
     import win32serviceutil
+    # threading for Windows service
+    import threading
 
 
 # Set some global variables for later
@@ -857,7 +859,7 @@ if __SYSTEM__ == 'nt':
             win32serviceutil.ServiceFramework.__init__(self, args)
             # handle WaitStop event tells the SCM to stop the service
             self.hWaitStop = win32event.CreateEvent(None, 0, 0, None)
-            self.running = False
+            self.running_event = threading.Event()
 
             # child process handles (Passive, Listener)
             self.p, self.l = None, None
@@ -916,40 +918,57 @@ if __SYSTEM__ == 'nt':
             logging.getLogger().setLevel(log_level)
 
         def SvcStop(self):
-            self.running = False
-            self.ReportServiceStatus(win32service.SERVICE_STOP_PENDING)
+            """
+            Stop the service
+            This triggers the stop event, which breaks the main loop
+            """
+            self.running_event.clear()
             win32event.SetEvent(self.hWaitStop) # set stop event for main thread
+
+        def SvcRun(self):
+            """
+            Start the service
+            We need to override this method to prevent it reporting NCPA as started before the processes are started
+            """
+            self.ReportServiceStatus(win32service.SERVICE_START_PENDING)
+            self.SvcDoRun()
+            # Once SvcDoRun returns, the service has stopped
+            self.ReportServiceStatus(win32service.SERVICE_STOPPED)
 
         def SvcDoRun(self):
             # log starting of service to windows event log
             servicemanager.LogMsg(servicemanager.EVENTLOG_INFORMATION_TYPE,
                                 servicemanager.PYS_SERVICE_STARTED,
                                 (self._svc_name_, ''))
-            self.running = True
+            self.running_event.set()
             self.main()
 
         def main(self):
-            self.ReportServiceStatus(win32service.SERVICE_START_PENDING)
-            # instantiate child processes
-            self.p, self.l = start_processes(self.options, self.config, self.has_error)
-            self.ReportServiceStatus(win32service.SERVICE_RUNNING)
+            try:
+                # instantiate child processes
+                self.p, self.l = start_processes(self.options, self.config, self.has_error)
+                self.ReportServiceStatus(win32service.SERVICE_RUNNING)
 
-            # wait for stop event
-            while self.running: # shouldn't loop, but just in case the event triggers without stop being called
-                win32event.WaitForSingleObject(self.hWaitStop, win32event.INFINITE)
-                time.sleep(1)
+                # wait for stop event
+                while self.running_event.is_set(): # shouldn't loop, but just in case the event triggers without stop being called
+                    result = win32event.WaitForSingleObject(self.hWaitStop, 1000)
+                    if result == win32event.WAIT_OBJECT_0:
+                        break
+                    time.sleep(0.1)
+            finally:
+                # kill/clean up child processes
+                if self.p:
+                    self.p.terminate()
+                    self.p.join() 
+                if self.l:
+                    self.l.terminate()
+                    self.l.join()
 
-            # kill/clean up child processes
-            self.p.terminate()
-            self.l.terminate()
-            self.p.join()
-            self.l.join()
-
-            # log stopping of service to windows event log
-            servicemanager.LogMsg(servicemanager.EVENTLOG_INFORMATION_TYPE,
-                                servicemanager.PYS_SERVICE_STOPPED,
-                                (self._svc_name_, ''))
-            self.ReportServiceStatus(win32service.SERVICE_STOPPED)
+                # log stopping of service to windows event log
+                servicemanager.LogMsg(servicemanager.EVENTLOG_INFORMATION_TYPE,
+                                    servicemanager.PYS_SERVICE_STOPPED,
+                                    (self._svc_name_, ''))
+                self.ReportServiceStatus(win32service.SERVICE_STOPPED)
 
 
 # --------------------------
